requirements.txt
Django==2.2.5
gmssl==3.2.1
Pillow==6.2.0
pycrypto==2.6.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
